+++ packages/js/src/MatomoTracker.ts
@@ -28,11 +28,11 @@ class MatomoTracker {
   static initialize({
     urlBase,
     siteId,
-    userId,
     trackerUrl,
     srcUrl,
     heartBeat,
     linkTracking = true,
+    configurations = {},
   }: UserOptions) {
     if (urlBase[urlBase.length - 1] !== '/') {
       urlBase = urlBase + '/'
@@ -48,9 +48,9 @@ class MatomoTracker {
         ])
         window._paq.push(['setSiteId', siteId])
 
-        if (userId) {
-          window._paq.push(['setUserId', userId])
-        }
+        Object.entries(configurations).forEach((entry) => {
+          window._paq.push([entry[0], entry[1]])
+        })
 
         // accurately measure the time spent on the last pageview of a visit
         if (!heartBeat || (heartBeat && heartBeat.active)) {