+++ packages/react/src/instance.ts
@@ -1,7 +1,7 @@
 import MatomoTracker from '@datapunt/matomo-tracker-js'
 import { InstanceParams } from './types'
 
-function createInstance(params: InstanceParams) {
+function createInstance(params: InstanceParams): MatomoTracker {
   return new MatomoTracker(params)
 }