+++ packages/react/src/utils/useOutboundClickListener.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { MatomoInstance } from '../types'
 
-const useOutboundClickListener = ({ trackLink }: MatomoInstance) => {
+const useOutboundClickListener = ({ trackLink }: MatomoInstance): void => {
   const handleOutboundClick = (event: MouseEvent) => {
     if (!event.target) {
       return
@@ -15,12 +15,14 @@ const useOutboundClickListener = ({ trackLink }: MatomoInstance) => {
       // Check if the click target differs from the current hostname, meaning it's external
       if (
         href &&
+        // eslint-disable-next-line @typescript-eslint/prefer-regexp-exec
         !href.match(
           new RegExp(
             `^(http://www.|https://www.|http://|https://)+(${window.location.hostname})`,
           ),
         )
       ) {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-call
         trackLink({ href })
       }
     }