+++ CHANGELOG.md
@@ -14,6 +14,11 @@ Prefix the change with one of these keywords:
 - _Fixed_: for any bug fixes.
 - _Security_: in case of vulnerabilities.
 
+## Unreleased
+
+- _Added_: include `configurations` in the `MatomoTracker.initialize` params
+- _Removed_: removed `userId` in the `MatomoTracker.initialize` params (user _configurations_ object instead)
+
 ## [0.1.5]
 
 - Added: [Exposed trackEvents function to useMatomo](https://github.com/Amsterdam/matomo-tracker/commit/79e96929c3fcde56434ec3ad82f24cb77d4225fd#diff-7f21e527da19e9a710bdcbb9a5387cbe)